Split test content generator to multiple modules
Split the test content generation script to multiple python modules, one
per content generator class

No refs
