+++ django/thunderstore/core/management/commands/create_test_data.py
@@ -1,426 +1,33 @@
-from abc import ABC
-from dataclasses import dataclass, field
-from typing import Any, Collection, Dict, List, Optional, OrderedDict, Type
+from typing import Dict, List, Optional, OrderedDict, Type
 
 from django.conf import settings
-from django.contrib.sites.models import Site
 from django.core.management.base import BaseCommand, CommandError
-from django.db.models import Model, QuerySet, signals
 
-from django_contracts.models import LegalContract, LegalContractVersion, PublishStatus
-from thunderstore.community.models import Community, CommunitySite, PackageListing
-from thunderstore.repository.factories import PackageVersionFactory
-from thunderstore.repository.models import Namespace, Package, PackageVersion, Team
-from thunderstore.utils.iterators import print_progress
-
-
-@dataclass
-class ContentPopulatorContext:
-    teams: Collection[Team] = field(default_factory=list)
-    packages: Collection[Package] = field(default_factory=list)
-    communities: Collection[Community] = field(default_factory=list)
-    contracts: Collection[LegalContract] = field(default_factory=list)
-
-    community_count: int = 0
-    dependency_count: int = 0
-    version_count: int = 0
-    team_count: int = 0
-    package_count: int = 0
-    contract_count: int = 0
-    contract_version_count: int = 0
-
-
-class ContentPopulator(ABC):
-    def populate(self, context: ContentPopulatorContext) -> None:
-        ...
-
-    def update_context(self, context) -> None:
-        ...
-
-    def clear(self) -> None:
-        ...
-
-
-class CommunityPopulator(ContentPopulator):
-    communities: Optional[List[Community]] = None
-    identifier_prefix = "test-community-"
-    name_prefix = "Test Community "
-
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating communities...")
-
-        last = last_obj.pk if (last_obj := Community.objects.last()) else 0
-
-        existing = list(
-            Community.objects.filter(name__startswith=self.name_prefix)[
-                : context.community_count
-            ]
-        )
-        remainder = context.community_count - len(existing)
-
-        self.communities = existing + [
-            Community.objects.create(
-                name=f"{self.name_prefix}{last + i}",
-                identifier=f"{self.identifier_prefix}{last + i}",
-            )
-            for i in print_progress(range(remainder), remainder)
-        ]
-
-    def update_context(self, context) -> None:
-        if self.communities is not None:
-            context.communities = self.communities
-        else:
-            context.communities = Community.objects.filter(
-                name__startswith=self.name_prefix
-            )[: context.community_count]
-
-    def clear(self) -> None:
-        print("Deleting existing test communities...")
-        Community.objects.filter(name__startswith=self.name_prefix).delete()
-
-
-class CommunitySitePopulator(ContentPopulator):
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating community sites...")
-
-        for community in context.communities:
-            community_site = CommunitySite.objects.filter(community=community).first()
-            if community_site:
-                continue
-            CommunitySite.objects.create(
-                community=community,
-                site=Site.objects.get_or_create(
-                    name="Thunderstore",
-                    domain=f"{community.identifier}.thunderstore.localhost",
-                )[0],
-            )
-
-    def update_context(self, context) -> None:
-        pass
-
-    def clear(self) -> None:
-        pass
-
-
-class TeamPopulator(ContentPopulator):
-    teams: Optional[List[Team]] = None
-    name_prefix = "Test_Team_"
-
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating teams...")
-
-        last = last_team.pk if (last_team := Team.objects.last()) else 0
-
-        existing = list(
-            Team.objects.filter(name__startswith=self.name_prefix)[: context.team_count]
-        )
-        remainder = context.team_count - len(existing)
-
-        self.teams = existing + [
-            Team.create(name=f"{self.name_prefix}{last + i}")
-            for i in print_progress(range(remainder), remainder)
-        ]
-
-    def update_context(self, context) -> None:
-        if self.teams is not None:
-            context.teams = self.teams
-        else:
-            context.teams = Team.objects.filter(name__startswith=self.name_prefix)[
-                : context.team_count
-            ]
-
-    def clear(self) -> None:
-        print("Deleting existing teams...")
-        Team.objects.all().delete()
-        print("Deleting existing namespaces...")
-        Namespace.objects.all().delete()
-
-
-class PackagePopulator(ContentPopulator):
-    packages: Optional[Collection[Package]] = None
-    name_prefix = "Test_Package_"
-
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating packages...")
-        packages = []
-
-        # Disabling signals to avoid spamming cache clear calls
-        # TODO: Implement a context manager for disabling cache refresh instead
-        signals.post_save.disconnect(Package.post_save, sender=Package)
-        signals.post_delete.disconnect(Package.post_delete, sender=Package)
-
-        for team in print_progress(context.teams, len(context.teams)):
-            existing = list(
-                team.owned_packages.filter(name__startswith=self.name_prefix)[
-                    : context.package_count
-                ]
-            )
-            offset = len(existing)
-            remainder = context.package_count - offset
-            team_packages = [
-                Package.objects.create(
-                    owner=team,
-                    name=f"{self.name_prefix}{i + offset}",
-                    namespace=team.get_namespace(),
-                )
-                for i in range(remainder)
-            ]
-            packages += team_packages
-
-        # Re-enabling previously disabled signals
-        signals.post_save.connect(Package.post_save, sender=Package)
-        signals.post_delete.connect(Package.post_delete, sender=Package)
-
-        self.packages = packages
-
-    def update_context(self, context) -> None:
-        if self.packages:
-            context.packages = self.packages
-        else:
-            context.packages = Package.objects.filter(
-                owner__in=context.teams,
-                name__startswith=self.name_prefix,
-            )
-
-    def clear(self) -> None:
-        print("Deleting existing packages...")
-        Package.objects.all().delete()
-
-
-class PackageVersionPopulator(ContentPopulator):
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating package versions...")
-
-        # Disabling signals to avoid spamming cache clear calls and other
-        # needless action
-        # TODO: Implement a context manager for disabling cache refresh instead
-        signals.post_save.disconnect(PackageVersion.post_save, sender=PackageVersion)
-        signals.post_delete.disconnect(
-            PackageVersion.post_delete, sender=PackageVersion
-        )
-        signals.post_save.disconnect(Package.post_save, sender=Package)
-        signals.post_delete.disconnect(Package.post_delete, sender=Package)
-
-        for i, package in print_progress(
-            enumerate(context.packages), len(context.packages)
-        ):
-            vercount = package.versions.count()
-            for vernum in range(context.version_count - vercount):
-                PackageVersionFactory.create(
-                    package=package,
-                    name=package.name,
-                    version_number=f"{vernum + vercount}.0.0",
-                    website_url="https://example.org",
-                    description=f"Example mod {i}",
-                    readme=f"# This is an example mod number {i}",
-                    changelog=f"# Example changelog for mod number {i}",
-                )
-
-            # Manually calling would-be signals once per package, as it doesn't
-            # actually make use of the sender param at all (and can be None)
-            package.handle_created_version(None)
-            package.handle_updated_version(None)
-
-        # Re-enabling previously disabled signals
-        signals.post_save.connect(PackageVersion.post_save, sender=PackageVersion)
-        signals.post_delete.connect(PackageVersion.post_delete, sender=PackageVersion)
-        signals.post_save.connect(Package.post_save, sender=Package)
-        signals.post_delete.connect(Package.post_delete, sender=Package)
-
-    def clear(self) -> None:
-        print("Deleting existing package versions...")
-        PackageVersion.objects.all().delete()
-
-
-class DependencyPopulator(ContentPopulator):
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Linking dependencies...")
-        PackageVersion.dependencies.through.objects.all().delete()
-        dependencies = [
-            x.latest.id for x in context.packages[: context.dependency_count]
-        ]
-        dependants = context.packages[context.dependency_count :]
-        for package in print_progress(dependants, len(dependants)):
-            package.latest.dependencies.set(dependencies)
-
-    def clear(self) -> None:
-        print("Deleting existing package dependency relations")
-        PackageVersion.dependencies.through.objects.all().delete()
-
-
-class ListingPopulator(ContentPopulator):
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print("Populating package listings")
-
-        # Disabling signals to avoid spamming cache clear calls
-        # TODO: Implement a context manager for disabling cache refresh instead
-        signals.post_save.disconnect(PackageListing.post_save, sender=PackageListing)
-        signals.post_delete.disconnect(
-            PackageListing.post_delete, sender=PackageListing
-        )
-
-        for i, package in print_progress(
-            enumerate(context.packages), len(context.packages)
-        ):
-            for community in context.communities:
-                package.get_or_create_package_listing(community)
-
-        # Re-enabling previously disabled signals
-        signals.post_save.connect(PackageListing.post_save, sender=PackageListing)
-        signals.post_delete.connect(PackageListing.post_delete, sender=PackageListing)
-
-    def clear(self) -> None:
-        print("Deleting existing package listings...")
-        PackageListing.objects.all().delete()
-
-
-class BaseContentPopulator(ContentPopulator):
-    model_cls: Type[Model] = None
-    name: str = None
-
-    def clear(self) -> None:
-        print(f"Deleting {self.name}...")
-        self.model_cls.objects.all().delete()
-
-    def get_last(self) -> int:
-        return last_obj.pk if (last_obj := self.model_cls.objects.last()) else 0
-
-    def update_context(self, context: ContentPopulatorContext) -> None:
-        pass
-
-    def set_context_objs(
-        self, context: ContentPopulatorContext, objs: List[Model]
-    ) -> None:
-        pass
-
-
-class LegalContractPopulator(BaseContentPopulator):
-    model_cls = LegalContract
-    objs: Optional[List[LegalContract]] = None
-    name = "legal contracts"
-    obj_name_prefix = "Test Contract "
-    obj_id_prefix = "test-contract-"
-    obj_id_field = "slug"
-
-    def get_existing(self) -> QuerySet:
-        return self.model_cls.objects.filter(
-            **{
-                f"{self.obj_id_field}__startswith": self.obj_id_prefix,
-            }
-        )
-
-    def create(self, index: int) -> LegalContract:
-        return self.model_cls.objects.create(**self.get_create_kwargs(index))
-
-    def get_create_kwargs(self, index: int) -> Dict[str, Any]:
-        return {
-            f"{self.obj_id_field}": f"{self.obj_id_prefix}{index}",
-            "title": f"{self.obj_name_prefix}{index}",
-        }
-
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print(f"Populating {self.name}...")
-
-        last = self.get_last()
-        existing = list(self.get_existing()[: self.get_context_count(context)])
-        remainder = self.get_context_count(context) - len(existing)
-
-        self.objs = existing + [
-            self.create(last + i) for i in print_progress(range(remainder), remainder)
-        ]
-        for obj in self.objs:
-            if obj.publish_status != PublishStatus.PUBLISHED:
-                obj.publish()
-
-    def update_context(self, context: ContentPopulatorContext) -> None:
-        if self.objs is not None:
-            self.set_context_objs(context, self.objs)
-        else:
-            self.set_context_objs(
-                context, self.get_existing()[: self.get_context_count(context)]
-            )
-
-    def get_context_count(self, context: ContentPopulatorContext):
-        return context.contract_count
-
-    def set_context_objs(
-        self, context: ContentPopulatorContext, objs: List[Any]
-    ) -> None:
-        context.contracts = objs
-
-
-LOREM_IPSUM = """
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-"""
-
-
-class LegalContractVersionPopulator(BaseContentPopulator):
-    model_cls = LegalContractVersion
-    name = "legal contract versions"
-    obj_id_field = "markdown_content"
-    obj_id_prefix = "## Test Contract "
-
-    def create(self, contract: LegalContract, index: int) -> LegalContractVersion:
-        return LegalContractVersion(
-            **{
-                "contract": contract,
-                self.obj_id_field: "\n".join(
-                    [
-                        f"{self.obj_id_prefix}{index} - Markdown",
-                        "",
-                        LOREM_IPSUM,
-                        "",
-                        "### Subtitle",
-                        "",
-                        LOREM_IPSUM,
-                    ]
-                ),
-                "html_content": "\n".join(
-                    [
-                        f"<h2>Test Contract {index} - HTML</h2>" "<br>",
-                        f"<p>{LOREM_IPSUM}</p>",
-                        "<br>",
-                        "<h3>Subtitle</h3>" "<br>",
-                        f"<p>{LOREM_IPSUM}</p>",
-                    ]
-                ),
-            }
-        )
-
-    def populate(self, context: ContentPopulatorContext) -> None:
-        print(f"Populating {self.name}...")
-
-        last = self.get_last()
-        objs = []
-        for contract in context.contracts:
-            existing = list(
-                contract.versions.filter(
-                    **{
-                        f"{self.obj_id_field}__startswith": self.obj_id_prefix,
-                        "contract": contract,
-                    }
-                )[: self.get_context_count(context)]
-            )
-            remainder = self.get_context_count(context) - len(existing)
-
-            objs.extend(
-                existing
-                + [
-                    self.create(contract, last + i)
-                    for i in print_progress(range(remainder), remainder)
-                ]
-            )
-
-        for obj in objs:
-            if obj.publish_status != PublishStatus.PUBLISHED:
-                obj.publish()
-
-    def get_context_count(self, context: ContentPopulatorContext):
-        return context.contract_version_count
-
+from thunderstore.core.management.commands.content.base import (
+    ContentPopulator,
+    ContentPopulatorContext,
+)
+from thunderstore.core.management.commands.content.community import CommunityPopulator
+from thunderstore.core.management.commands.content.community_site import (
+    CommunitySitePopulator,
+)
+from thunderstore.core.management.commands.content.contract import (
+    LegalContractPopulator,
+)
+from thunderstore.core.management.commands.content.contract_version import (
+    LegalContractVersionPopulator,
+)
+from thunderstore.core.management.commands.content.dependencies import (
+    DependencyPopulator,
+)
+from thunderstore.core.management.commands.content.package import PackagePopulator
+from thunderstore.core.management.commands.content.package_listing import (
+    ListingPopulator,
+)
+from thunderstore.core.management.commands.content.package_version import (
+    PackageVersionPopulator,
+)
+from thunderstore.core.management.commands.content.team import TeamPopulator
 
 # In generation order; clearing order is inverted
 CONTENT_POPULATORS: Dict[str, Type[ContentPopulator]] = OrderedDict[
